requirements.txt
astroid==2.11.5
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dill==0.3.5.1
Flask==2.1.2
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
Pillow==9.1.1 
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.14.1
requests==2.27.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
